Solved merge conflicts
* Solved merge conflicts, left from previous commit. Next time, do a
'git diff' before you commit, and if there are any symbols like
'>>>>>>', it means you have to make a choice between two versions of
block of code. Remove the one you don't need, as well as the '>>>>'
lines, and commit again :)
* Converted tabs to spaces. For python, it's very important to stick to
either tabs everywhere, or 4 spaces. if you mix the two, sometimes the
program will not work, as the indentation you see won't match what the
interpreter sees. Had some nasty bugs with that in the past. I recomment
to use PyCharm for python development, as it shows you when it's wrong.
* Probably changed Windows line endings to Linux (CRLF -> LF). Windows
uses '\r\n' for newline, while linux uses '\n'. This can sometimes cause
issues when attemptign to run the script. Not sure how to best avoid it.